requirements.txt
# MIRRALISM V5 必要パッケージ
psycopg2-binary==2.9.7
chromadb==0.4.15
sentence-transformers==2.2.2
openai==1.3.5
python-dotenv==1.0.0
pandas==2.1.3
numpy==1.25.2
markdown==3.5.1